Add files via upload
Problem background
For the past few months, Airbnb has seen a major decline in revenue. Now that the restrictions have started lifting and people have started to travel more, Airbnb wants to make sure that it is fully prepared for this change.

End Objective
Analyze the dataset consisting of various Airbnb listings in New York.